backend/requirements.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.9
google-auth>=2.41
google-auth-oauthlib>=1.2
Pillow>=12.1
